+++ extensions/measurement-tracking/src/viewports/TrackedCornerstoneViewport.tsx
@@ -141,7 +141,9 @@ function TrackedCornerstoneViewport(props) {
           // Only send the tracked measurements event for the active viewport to avoid
           // sending it more than once.
           if (viewportId === activeViewportId) {
-            const { referenceStudyUID: StudyInstanceUID, referenceSeriesUID: SeriesInstanceUID } =
+            const { referenceStudyUID: StudyInstanceUID,
+              referenceSeriesUID: SeriesInstanceUID,
+              uid: measurementId } =
               measurement;
 
             sendTrackedMeasurementsEvent('SET_DIRTY', { SeriesInstanceUID });
@@ -149,6 +151,7 @@ function TrackedCornerstoneViewport(props) {
               viewportId,
               StudyInstanceUID,
               SeriesInstanceUID,
+              measurementId
             });
           }
         }).unsubscribe