+++ platform/ui/src/components/index.js
@@ -83,6 +83,7 @@ import PanelSection from './PanelSection';
 import AdvancedToolbox from './AdvancedToolbox';
 import InputDoubleRange from './InputDoubleRange';
 import LegacyButtonGroup from './LegacyButtonGroup';
+import LabellingFlow from './Labelling';
 import SwitchButton, { SwitchLabelLocation } from './SwitchButton';
 import * as AllInOneMenu from './AllInOneMenu';
 import ViewportActionArrows from './ViewportActionArrows';
@@ -92,7 +93,6 @@ import { ToolSettings } from './AdvancedToolbox';
 import { Toolbox } from './Toolbox';
 import InvestigationalUseDialog from './InvestigationalUseDialog';
 import MeasurementItem from './MeasurementTable/MeasurementItem';
-
 import LayoutPreset from './LayoutPreset';
 
 export {
@@ -188,6 +188,7 @@ export {
   WindowLevel,
   WindowLevelMenuItem,
   InputDoubleRange,
+  LabellingFlow,
   HeaderPatientInfo,
   LayoutPreset,
   LegacySplitButton,